+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.33.2 is now available from:
 
   <https://download.bitcoinabc.org/0.33.2/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - A new Chronik endpoint `/script/batch/utxos` has been added to fetch the
+   utxos of several scripts in a single request.